dev-requirements.txt
bcrypt==4.1.3
pycodestyle==2.6.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-decouple==3.8
sqlparse==0.5.0
